Working on a simple catch-too-much LL(1) grammar for XY.
Ideally I will be able to use the parser generated as a fast pass
through a file to pull out relevant info, when, for e.g. including
other modules. It would help to not have the same level of error
checking / etc. Looks like this grammar also induces some bug so I will
test that on mccrae.